refactor: Enum 클래스 패키지 이동
domain, validator, view의 constants 패키지로 이동
domain 내의 Enum 클래스는 역할에 따라 분리
