+++ docs/README.md
@@ -1,8 +1,10 @@
 # 목차
 1. [크리스마스 프로모션 개요](#🎄-크리스마스-프로모션-개요)
 2. [기능 목록](#📋-기능-목록)
-3. [예외 상황](#⚠️-예외 상황)
+3. [예외 상황](#⚠️-예외-상황)
+4. [패키지 구조 설명](#📦-패키지-구조-설명)
 
+<br>
 
 # 🎄 크리스마스 프로모션 개요
 
@@ -99,3 +101,7 @@
 - [x] 메뉴 형식이 ‘$메뉴-$개수’가 아닐 경우
 - [x] 중복 메뉴를 입력한 경우
 - [x] 음료만 주문한 경우
+
+<br>
+
+# 📦 패키지 구조 설명
