+++ src/main/java/christmas/domain/contants/eventInfo/SpecialEventDate.java
@@ -1,7 +1,6 @@
-package christmas.domain.contants;
+package christmas.domain.contants.eventInfo;
 
 import christmas.domain.VisitDate;
-import christmas.util.Calculator;
 
 import java.util.Arrays;
