+++ src/main/java/christmas/util/Parser.java
@@ -1,7 +1,7 @@
 package christmas.util;
 
-import christmas.domain.contants.InputSetting;
-import christmas.domain.contants.ReportSetting;
+import christmas.domain.contants.order.OrderInputSetting;
+import christmas.domain.contants.benefitReport.ReportSetting;
 
 import java.text.NumberFormat;
 import java.util.Arrays;
@@ -13,11 +13,11 @@ public class Parser {
     }
 
     public static List<String> toOrderList(String playerInput) {
-        return Arrays.stream(playerInput.split(InputSetting.ORDER_DELIMITER.getSetting())).toList();
+        return Arrays.stream(playerInput.split(OrderInputSetting.ORDER_DELIMITER.getSetting())).toList();
     }
 
     public static String[] toMenuAndCountArray(String order) {
-        return order.split(InputSetting.MENU_COUNT_DELIMITER.getSetting());
+        return order.split(OrderInputSetting.MENU_COUNT_DELIMITER.getSetting());
     }
 
     public static String toThousandUnitMoney(int money) {
